requirements.txt
google-cloud-bigquery
dbt-bigquery
dbt-postgres
python-dotenv
psycopg2
pandas
numpy
seaborn